fix: correcion de campos en las migraciones y cambio de mensajes de ingles a espaniol

+++ database/migrations/2023_05_02_140630_create_products_table.php
@@ -20,10 +20,10 @@ return new class extends Migration
             $table->string('code');
             //$table->string('product_barcode_symbology')->nullable();
             $table->integer('quantity');
-            $table->integer('buying_price')->comment('Buying Price');
-            $table->integer('selling_price')->comment('Selling Price');
+            $table->double('buying_price')->comment('Buying Price');
+            $table->double('selling_price')->comment('Selling Price');
             $table->integer('quantity_alert');
-            $table->integer('tax')->nullable();
+            $table->double('tax')->nullable();
             $table->tinyInteger('tax_type')->nullable();
             $table->text('notes')->nullable();
