+++ database/migrations/2023_05_04_084646_create_order_details_table.php
@@ -23,8 +23,8 @@ return new class extends Migration
                 ->cascadeOnDelete();
 
             $table->integer('quantity');
-            $table->integer('unitcost');
-            $table->integer('total');
+            $table->double('unitcost');
+            $table->double('total');
             $table->timestamps();
         });
     }
